Version C
+ Version C

The version C has classes for layers and neurons, as well as better training, Version A and Version B are unusable, as they are somehow broken, it's easier to just make a new neural network from scratch instead of trying to figure it out. And version C has custom input/layer/hiddenNodes/outputs sizes. As well as more parameters to play with.